requirements.txt
Flask==1.1.2
gunicorn==20.0.4
PyYAML==5.4.1
requests==2.25.1